:rocket: (NextAuth): Upgraded to NextAuth v5 (#55)
* :rocket: (NextAuth): Upgraded to NextAuth v5

Using nextauth v5, added a middleware file, necessary nextauth changes and some UI enhancements to tell the user why they can't access the /calendar page

* :sparkles: (Hourly Tasks): Added Hourly Task Functionality

A complete overhaul of the way Tasks are rendered on the main screen. In DayView and WeekView, tasks are now ordered by the hour. The user can drag and drop
tasks anywhere now. A few bugs introduced with this but nothing that can't be ironed out

* :sparkles: (Calendar page): Calendar view buttons

Day view, Week view and Month views updates with next and previous buttons based on the Calendar view. Added a border to the Month view